requirements.txt
django == 2.2.6
djangorestframework == 3.10.3
django-cors-headers == 3.1.1